requirements.txt
annotated-types==0.7.0
anyio==4.4.0
app==0.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
cryptography==44.0.0
DateTime==5.5
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.3
django-request==1.6.3
executing==2.0.1
Faker==33.1.0
fanstatic==1.4
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Admin==1.6.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
fuzzywuzzy==0.18.0
gevent==25.4.1
geventhttpclient==2.3.3
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet>=3.2.0
grpcio==1.67.1
grpcio-status==1.67.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
js==1.0
json5==0.9.25
jsonify==0.5
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
legacy-cgi==2.6.1
Levenshtein==0.26.1
load-dotenv==0.1.0
locust==2.35.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.1.0
MyApplication==0.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.1
python-json-logger==2.0.7
python-Levenshtein==0.26.1
python-socketio==5.12.0
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
RapidFuzz==3.10.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
selenium==4.31.0
Send2Trash==1.8.3
setuptools==75.6.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
unicode==2.9
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
utils==1.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.8.0
Werkzeug==3.1.0
widgetsnbextension==4.0.13
wsproto==1.2.0
WTForms==3.2.1
zope.event==5.0
zope.interface==7.2
